Adding ability to allow HTML emails
Adding a new Admin Settings option to "Allow HTML Emails". Checking this will
allow ooSTicket to recieve HTML emails and display as best it can. The HTML
Emails can be an attachment or inline. There is some issue with mails sent
from GMail, but this is not a show stoppe for its usage for us.