Ability to View HTML Emails
Adding the functionality to view HTML formatted emails. This does its
best to view HTML emails as they were intended. It is not perfect, and
there is an issue with mails sent from GMail*, but for our purposes it
is good enough.

New Admin setting to turn the viewing of HTML formatted emails on or
off.

Slight database change.

* Gmail bug is that if a user send from gmail with an inline attachment
  then the whole email is stipped. Obviously this is not very good, bu
  the number of Gmail users which will send us emails like this are
  minimal, so it's not stopping us go live.